tools: parse_*: add support for own custom HID extension with offsets
The extension proposed here is to had several vendor usage pages, each
of them prefixed by 0xbe, and the LSB is a current normalized usage page.

For instance, the Digitizer Usage Page (0x000d) becomes 0xbe0d with this
vendor collection. Once we have an usage in this vendor page, we also
split the 2 bytes of the usage in 2:
 - the MSB reflects the offset of the value within the usage
 - the LSB reflects the actual usage in the original page

So if we encounter a usage page of 0xbe01 with a usage 0x0930, that mean
that we have the usage page Generic Desktop with an offset, and the usage
is X which will be shifted by 9 to the left (X << 9).

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>