hid-replay: reply error when receiving GET_REPORT while in probe()
If the driver needs to access some data from the device while in probe(),
it may issue some GET_REPORT HID commands, which currently timeout.

Instead of waiting a few seconds, we can tell the driver we don't support
those directly and let it continue the initialization.

Currently, if a GET_REPORT request comes after the device has been probed,
we do not handle those.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>