hid-replay: Wait for all devices and error/exit if this fails
The existing wait loop in `hid_replay_wait_opened()` exits as soon as a
single device has been sucessfully created. Really, however, we should
wait for *all* devices to reach this state before continuing in case
an error has occurred something (e.g. the kernel failed to `probe()`
the device). Now that we can sensibly rely on the timeout mechanism
in `hid_replay_get_one_event()` (without being flooded by constant
'0' results if the user happened to hit a key) we can allow a short
amount of time for all devices to get ready and error out with a
helpful message if any are apparently having issues.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>