hid-replay: Remove STDIN from poll() list to enable sensible device waiting
The `hid_replay_get_one_event` currently returns '0' both when the timeout
has been exceeded, as well as when the only ready fd is STDIN. This makes
checking for "real" timeouts difficult since you have to ignore the case
where STDIN is ready. It appears the only reason STDIN is even part of the
fd list is to make the "Hit enter" logic work.

This patch removes STDIN from the list of devices we poll() on so that '0'
is only returned in the case of a real timeout. To make the "Hit enter"
logic work without listening on STDIN, we repeatedly call the
`hid_replay_get_one_event` function in a loop until a small-ish timeout
occurs and we can be reasonably sure all events have been drained.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>