Always define _GNU_SOURCE
Make this the business of the build system, not of individual files. There's
no good reason to have it defined in one file but not the other one, it's just
prone to trigger bugs.

Because the simple approach for automake (AC_USE_SYSTEM_EXTENSIONS) writes it
to config.h, we need to do the same thing in meson. Once automake is dropped
we can just add it as global argument.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>